refactor: 커스텀 훅 폴더 구조 수정
- lib 폴더 내부에 있는 폴더 외부로 분리
