+++ hooks/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { default as useCardNumber } from "./hooks/useCardNumber";
-export { default as useCvcNumber } from "./hooks/useCvcNumber";
-export { default as useExpirationDate } from "./hooks/useExpirationDate";
-export { default as usePassword } from "./hooks/usePassword";
+export { default as useCardNumber } from "../hooks/useCardNumber";
+export { default as useCvcNumber } from "../hooks/useCvcNumber";
+export { default as useExpirationDate } from "../hooks/useExpirationDate";
+export { default as usePassword } from "../hooks/usePassword";
